Update External_data_sources.md
Minor typo fixes.

I'm wondering if this file should include datestamps for when the versions of the data sources were loaded--e.g., "The 2012 version of the ModelSEED database was loaded into KBase on [date]". It also might be useful to give some idea of the update schedule. Will the 2012 version be used in perpetuity? Some other data sources are presumably updated in KBase more frequently?