Update How_to_add_a_new_data_type.md
Added title, made some minor wording/punctuation improvements.

Added FIXME to that places that need work, to make it harder to forget to update them.

Comment on line 417: "From solrconfig.xml you need to modify line 106 to point to the correct directory for the index data of your type." If there's a new release of solrconfig.xml, it might not be line 106 anymore, so you should include that line here (or at least say which variable you need to set/modify). (Similarly for the other line numbers mentioned here--describe what's on the lines. No one is going to think to come and update this document if the SOLR installation is updated and the line numbers change!)

Line 429: Does Solr use the term "ingest"? If not, I would change this to "input" (but if "ingest" is their term for it, then it's ok to leave it).