Added a section on User acceptance tests
It could be optionally added under Testing section. However, user acceptance tests in KBase have a fairly wide scope and target many aspects of the platform - functionality, documentation, data sets, website, outreach and user engagement. Therefore, adding it as a separate section in the end of the document. 

Note: Thanks to Bob and Doreen for helping shape the beta testing process at KBase.